Ditch util.py
