+++ integration-testing/rnode_testing/wait.py
@@ -2,7 +2,6 @@ import logging
 import re
 import pytest
 import time
-from rnode_testing.util import log_box
 
 
 def wait_for(condition, timeout, error_message):
